refactor(glue): split abi module.
