+++ crates/glue/src/abi/alloc.rs
@@ -1,0 +1,11 @@
+#[no_mangle]
+pub extern "C" fn alloc(len: usize) -> *mut u8 {
+    let layout = std::alloc::Layout::from_size_align(len, 1).unwrap();
+    unsafe { std::alloc::alloc(layout) }
+}
+
+#[no_mangle]
+pub extern "C" fn dealloc(ptr: *mut u8, len: usize) {
+    let layout = std::alloc::Layout::from_size_align(len, 1).unwrap();
+    unsafe { std::alloc::dealloc(ptr, layout) }
+}
